Extract method to new class
